task_Management/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-ckeditor==6.2.0
django-crontab==0.7.1
django-js-asset==1.2.2
django-last-seen==0.3
django-session-timeout==0.1.0
django-tinymce==3.4.0
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
gunicorn==20.1.0
mysqlclient==2.0.3
Pillow==8.3.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
xlwt==1.3.0
